+++ packages/clean_code_data/lib/src/services/app_installation_service.dart
@@ -53,7 +53,7 @@ class AppInstallationServiceImpl
       return InstallationEntity(
         installationId: encodedInstallationId,
         appName: appName,
-        appVersion: appInfo.version,
+        appVersion: appInfo.versionOnly,
         appIdentifier: appIdentifier,
         channels: [],
         gcmSenderId: gcmSenderId,