+++ packages/clean_code_data/lib/src/use_cases/encrypt_server_public_key_use_case.dart
@@ -17,6 +17,6 @@ class EncryptServerPublicKeyUseCaseImpl
     final String key = _serverRSAPublicKeyBase64;
     if (key.isEmpty) return data;
     final String publicKey = utf8.decode(base64.decode(key));
-    return _rsaEncrypterUseCase.encrypt(publicKey: publicKey, data: data);
+    return await _rsaEncrypterUseCase.encrypt(publicKey: publicKey, data: data);
   }
 }